requirements.txt
# torch==2.1.0
# torchvision==0.16.0
opencv-python==4.10.0
PyYAML==6.0
monai==1.3.2
tensorboard==2.15.0
gudhi==3.10.1
numpy==1.26.0
albumentations==1.4.14
matplotlib